Consolidated global/index css files, resolved imports in toast components
